--- src/styles/home.scss
@@ -1,120 +1,0 @@
-@keyframes breath {
-    0% {
-        opacity: 0;
-    }
-
-    100% {
-        opacity: 1;
-    }
-}
-
-.Home {
-    height: 100vh;
-    max-width: 100%;
-    background-color: rgb(20, 20, 20);
-
-    .heading {
-        height: 90%;
-        width: 90%;
-        display: flex;
-        justify-content: center;
-        align-items: start;
-        padding-left: 50px;
-        flex-direction: column;
-        gap: 10px;
-
-        span {
-            color: rgb(171, 166, 166);
-            font-family: sans-serif;
-            font-size: 34px;
-            user-select: none;
-            height: 50px;
-        }
-
-        .popout {
-            h1 {
-                cursor: pointer;
-                margin-right: 20px;
-                display: inline-block;
-                font-family: sans-serif;
-                color: white;
-                font-weight: 700;
-                font-size: 70px;
-                user-select: none;
-            }
-        }
-        p {
-            font-family: sans-serif;
-            color: rgb(171, 166, 166);
-            font-size: 25px;
-            user-select: none;
-        }
-    }
-}
-
-@media screen and (max-width: 880px) {
-    .Home {
-        .heading {
-            padding-left: 20px;
-
-            .popout {
-                h1 {
-                    font-size: 60px;
-                }
-            }
-
-            span {
-                font-size: 30px;
-            }
-        }
-
-    }
-}
-
-@media screen and (max-width: 573px) {
-    .Home {
-        .heading {
-            .popout {
-                width: 300px;
-                overflow-wrap: break-word;
-                word-wrap: break-word;
-                word-break: break-word;
-            }
-
-            span {
-                font-size: 30px;
-            }
-        }
-    }
-}
-
-@media screen and (max-width: 401px) {
-    .Home {
-        @keyframes moveUp {
-            0% {
-                transform: translateY(150px);
-            }
-
-            100% {
-                transform: translateY(0);
-            }
-        }
-    }
-}
-
-@media screen and (max-width: 351px) {
-    .Home {
-        .heading {
-            padding-left: 17px;
-            
-            .popout {
-                h1 {
-                    font-size: 55px;
-                }
-            }
-            span {
-                font-size: 27px;
-            }
-        }
-    }
-}
