--- src/styles/project.scss
@@ -1,79 +1,0 @@
-.Project {
-    max-height: 600px;
-    overflow: hidden;
-    max-width: 370px;
-    min-height: 300px;
-    background-color: #2d2c2c;
-    box-shadow: 0 2px 4px -1px rgb(0 0 0 / 20%), 0 4px 5px 0 rgb(0 0 0 / 14%), 0 1px 10px 0 rgb(0 0 0 / 12%);
-    border-top: 2px solid var(--color);
-    transition: transform .2s ease-in-out;
-    padding-bottom: 50px;
-
-    h2,
-    p {
-        font-family: sans-serif;
-    }
-
-    h2 {
-        font-size: 30px;
-        padding-top: 30px;
-        padding-left: 30px;
-        padding-right: 30px;
-        color: white;
-        transition: 0.5s;
-    }
-
-    p {
-        font-size: 20px;
-        padding-top: 10px;
-        padding-left: 30px;
-        padding-right: 30px;
-        color: #A7A7A7;
-    }
-
-    p.Language {
-        color: var(--color);
-        font-size: 17px;
-        padding-top: 10px;
-        letter-spacing: 0.5px;
-        padding-left: 30px;
-        padding-right: 30px;
-    }
-
-    button.raise:hover,
-    button.raise:focus {
-        box-shadow: 0 0.5em 0.5em -0.4em;
-        transform: translateY(-0.25em);
-    }
-
-    button {
-        cursor: pointer;
-        color: white;
-        transition: 0.25s;
-        height: 50px;
-        width: 130px;
-        margin-top: 30px;
-        margin-left: 30px;
-        margin-right: 30px;
-        font-size: 15px;
-        background: none;
-        font-family: sans-serif;
-        border: 2px solid white;
-
-        &:hover,
-        &:focus {
-            border-color: var(--color);
-            color: var(--color);
-        }
-    }
-}
-
-.Project:hover {
-    transform: scale(1.02);
-}
-
-@media screen and (max-width: 500px) {
-    .Project:hover {
-        transform: none;
-    }
-}
