--- src/styles/projects.scss
@@ -1,61 +1,0 @@
-.Projects {
-    scroll-snap-align: start;
-    min-height: 135vh;
-    margin: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    gap: 10px;
-
-    .heading h1 {
-        text-align: center;
-        margin: 0 auto;
-        padding-top: 40px;
-        font: 300 45px;
-        font-family: sans-serif;
-        color: #fff;
-        text-transform: uppercase;
-        letter-spacing: 6px;
-    }
-
-    .row-1 {
-        padding-top: 40px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        flex-direction: row;
-        gap: 10px;
-    }
-
-    .row-2 {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        flex-direction: row;
-        gap: 10px;
-        padding-bottom: 5px;
-    }
-}
-
-@media screen and (max-width: 1000px) {
-    .Projects {
-        flex-direction: column;
-
-        .row-1 {
-            flex-direction: column;
-            margin-top: 0px;
-        }
-
-        .row-2 {
-            flex-direction: column;
-        }
-    }
-}
-
-@media screen and (max-width: 400px) {
-
-    .Projects {
-        padding: 0px 10px 0px 10px;
-    }
-}
